Plot method for summary.estimate_scr_by objects

Plot method for summary.estimate_scr_by objects

Usage

# S3 method for class 'summary.estimate_scr_by'
autoplot(
  object,
  xvar,
  fill_var,
  alpha = 0.7,
  CIs = FALSE,
  title = "Seroconversion Rate by Group",
  xlab = "Seroconversion rate per 1000 person-years",
  ylab = "Stratification Variable",
  fill_palette = NULL,
  ...
)

Arguments

object

a summary.seroincidence.by object (generated by applying the summary() method to the output of estimate_scr_by()).

xvar

the name of a stratifying variable in object (e.g., "ageCat")

fill_var

the name of the variable used for fill color (e.g., "Country")

alpha

transparency for the bars (1 = no transparency, 0 = fully transparent)

CIs

logical, if TRUE, add horizontal CI error bars

title

a title for the final plot

xlab

a label for the x axis of the final plot

fill_palette

define the color palette for bar fill

...

unused

yvar

the name of a second stratifying variable in object (e.g., "catchment")

fill_lab

fill label

Value

a ggplot2::ggplot() object

Examples

library(dplyr)
library(ggplot2)

xs_data <-
  sees_pop_data_pk_100

curve <-
  typhoid_curves_nostrat_100 %>%
  filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG"))

noise <-
  example_noise_params_pk

est_catchment_age <- estimate_scr_by(
strata = c("catchment", "ageCat"),
pop_data = xs_data,
curve_params = curve,
curve_strata_varnames = NULL,
noise_params = noise,
noise_strata_varnames = "catchment",
antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
num_cores = 8 # Allow for parallel processing to decrease run time
)
#> Warning: `noise_params` is missing all strata variables and will be used unstratified.
#> ℹ To avoid this warning, specify the desired set of stratifying variables in
#>   the `curve_strata_varnames` and `noise_strata_varnames` arguments to
#>   `estimate_scr_by()`.
#> This computer appears to have 4 cores available. `estimate_scr_by()` has
#> reduced its `num_cores` argument to 3 to avoid destabilizing the computer.

est_catchment_age_sum <- summary(est_catchment_age)

autoplot.summary.estimate_scr_by(
  est_catchment_age_sum,
  "ageCat",
  fill_var = "catchment"
)
#> Error in autoplot.summary.estimate_scr_by(est_catchment_age_sum, "ageCat",     fill_var = "catchment"): could not find function "autoplot.summary.estimate_scr_by"