Plot method for summary.seroincidence.by objects

Plot method for summary.seroincidence.by objects

Usage

# S3 method for class 'summary.seroincidence.by'
autoplot(object, xvar, alpha = 0.7, shape = 1, width = 0.001, ...)

Arguments

object

a summary.seroincidence.by object (generated by applying the summary() method to the output of est.incidence.by()).

xvar

the name of a stratifying variable in object

alpha

transparency for the points in the graph (1 = no transparency, 0 = fully transparent)

shape

shape argument for geom_point()

width

width for jitter

...

unused

Value

a ggplot2::ggplot() object

Examples

library(dplyr)
library(ggplot2)

xs_data <-
  sees_pop_data_pk_100

curve <-
  typhoid_curves_nostrat_100 %>%
  filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG"))

noise <-
  example_noise_params_pk

est2 <- est.incidence.by(
  strata = c("catchment"),
  pop_data = xs_data,
  curve_params = curve,
  noise_params = noise,
  antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
  #num_cores = 8 #Allow for parallel processing to decrease run time
)
#> Warning: curve_params is missing all strata variables, and will be used unstratified.
#> 
#> To avoid this warning, specify the desired set of stratifying variables in the `curve_strata_varnames` and `noise_strata_varnames` arguments to `est.incidence.by()`.
#> Warning: noise_params is missing all strata variables, and will be used unstratified.
#> 
#> To avoid this warning, specify the desired set of stratifying variables in the `curve_strata_varnames` and `noise_strata_varnames` arguments to `est.incidence.by()`.
#> Error in (function (pop_data, curve_params, noise_params, antigen_isos = pop_data$antigen_iso %>%     unique(), lambda_start = 0.1, stepmin = 1e-08, stepmax = 3,     verbose = FALSE, build_graph = FALSE, print_graph = build_graph &         verbose, ...) {    if (verbose > 1) {        message("inputs to `est.incidence()`:")        print(environment() %>% as.list())    }    .errorCheck(data = pop_data, antigen_isos = antigen_isos,         curve_params = curve_params)    pop_data <- pop_data %>% dplyr::filter(.data$antigen_iso %in%         antigen_isos) %>% dplyr::select(pop_data %>% get_value_var(),         pop_data %>% get_age_var(), "antigen_iso") %>% filter(if_all(everything(),         ~!is.na(.x)))    curve_params <- curve_params %>% ungroup() %>% dplyr::mutate(alpha = .data$alpha *         365.25, d = .data$r - 1) %>% dplyr::filter(.data$antigen_iso %in%         antigen_isos) %>% dplyr::select("y1", "alpha", "d", "antigen_iso") %>%         droplevels()    noise_params <- noise_params %>% dplyr::filter(.data$antigen_iso %in%         antigen_isos) %>% droplevels()    if (nrow(pop_data) == 0) {        stop("No data provided.")    }    if (verbose) {        message("nrow(curve_params) = ", nrow(curve_params))    }    if (nrow(noise_params) != length(antigen_isos)) {        stop("too many rows of noise parameters.")    }    pop_data <- pop_data %>% split(~antigen_iso)    curve_params <- curve_params %>% split(~antigen_iso)    noise_params <- noise_params %>% split(~antigen_iso)    res <- .nll(pop_data = pop_data, log.lambda = log(lambda_start),         antigen_isos = antigen_isos, curve_params = curve_params,         noise_params = noise_params, verbose = verbose, ...)    if (is.na(res)) {        warning("Could not calculate the log-likelihood with starting parameter value.")        return(NULL)    }    if (verbose) {        message("Initial negative log-likelihood: ", res)    }    if (build_graph) {        if (verbose)             message("building likelihood graph")        graph <- graph_loglik(highlight_points = lambda_start,             highlight_point_names = "lambda_start", pop_data = pop_data,             antigen_isos = antigen_isos, curve_params = curve_params,             noise_params = noise_params)        if (print_graph) {            print(graph + ggplot2::scale_x_log10(labels = scales::label_comma()))        }    }    else {        graph <- NULL    }    if (verbose)         message("about to call `nlm()`")    time <- {        fit <- nlm(f = .nll, p = log(lambda_start), pop_data = pop_data,             antigen_isos = antigen_isos, curve_params = curve_params,             noise_params = noise_params, hessian = TRUE, stepmax = stepmax,             steptol = stepmin, verbose = verbose, print.level = ifelse(verbose,                 2, 0), ...)    } %>% system.time()    code_text <- nlm_exit_codes[fit$code]    message1 <- "\n`nlm()` completed with the following convergence code:\n"    if (fit$code %in% 3:5) {        warning("`nlm()` may not have reached the maximum likelihood estimate.",             message1, code_text)    }    if (verbose) {        message("\nElapsed time: ")        print(time)    }    if (build_graph) {        graph <- graph %>% add_point_to_graph(fit = fit, pop_data = pop_data,             antigen_isos = antigen_isos, curve_params = curve_params,             noise_params = noise_params)        if (print_graph) {            print(graph + ggplot2::scale_x_log10(labels = scales::label_comma()))        }    }    fit <- fit %>% structure(class = union("seroincidence", class(fit)),         lambda_start = lambda_start, antigen_isos = antigen_isos,         ll_graph = graph)    return(fit)})(pop_data = structure(list(value = c(5.686082087, 1.228468773, 1.08478637, 1.429250766, 3.061925394, 3.10341076, 0.030783008, 6.39056873, 0.240410018, 3.454038997, 0.583069732, 1.203041652, 14.45068664, 13.67451493, 9.75485189, 1.854591567, 13.47695391, 3.915196534, 0.581800446, 0.464968826, 0.804765886, 0.170707915, 1.37100964, 3.117010281, 0.631406531, 0.457043221, 2.425332461, 1.706190158, 3.637831218, 69.89535911, 2.34926128, 2.421981965, 5.959094924, 1.208630376, 21.66881837, 1.980702589, 6.947747503, 0.326101424, 1.746848591, 5.743808809, 66.36882517, 8.592622525, 9.559545113, 1.463210702, 1.675724638, 2.001794061, 1.963263676, 0.144000922, 2.640976902, 2.639804308, 14.97160989, 0.581763723, 1.597643699, 4.211346348, 3.001945296, 0.216909126, 0.956437161, 9.571740156, 1.354560173, 1.019806649, 33.61421131, 0.234319736, 1.969896519, 1.97546336, 0.369090624, 3.81341291, 6.564569311, 7.044133815, 2.419830229, 5.613969982, 0.677328316, 0.430965682, 0.982963924, 0.3471668, 0.678930651, 1.373231907, 1.358157942, 1.607816553, 0.661695355, 3.73191269, 2.028698664, 3.084281709, 16.42857143, 4.100071759, 19.97064837, 8.296130952, 14.04425829, 6.958705357, 3.523065476, 9.819568452, 0.252788832, 0.875601873, 1.08911496, 6.30952381, 5.780099594, 5.30809886, 1.083906829, 0.807305453, 13.32860629, 0.53330289, 1.402155499, 6.947698519, 4.527529762, 1.970971466, 4.871499117, 2.123841957), age = c(18, 7.3, 2.6, 3.9, 13, 14.4, 3.4, 22, 4, 14.8, 15.1, 13, 15.6, 18, 18.7, 7.5, 23, 4, 2.3, 9.8, 21, 14.7, 4, 18.1, 11.2, 3.8, 11.5, 17.2, 17, 12, 11, 16, 5, 12, 20, 14, 7.3, 2.5, 10.5, 4.4, 11.2, 13.4, 14, 14.6, 19.2, 10, 5, 4.9, 9.2, 15.8, 18, 21, 6.6, 18, 7.3, 2.6, 3.9, 13, 14.4, 3.4, 22, 4, 14.8, 15.1, 13, 15.6, 18, 18.7, 7.5, 23, 4, 2.3, 9.8, 21, 14.7, 4, 18.1, 11.2, 3.8, 11.5, 17.2, 17, 12, 11, 16, 5, 12, 20, 14, 7.3, 2.5, 10.5, 4.4, 11.2, 13.4, 14, 14.6, 19.2, 10, 5, 4.9, 9.2, 15.8, 18, 21, 6.6), antigen_iso = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L), levels = c("HlyE_IgA", "HlyE_IgG"), class = "factor")), row.names = c(NA, -106L), class = c("pop_data", "tbl_df", "tbl", "data.frame"), antigen_isos = structure(1:2, levels = c("HlyE_IgA", "HlyE_IgG"), class = "factor"), age_var = "age", value_var = "value", id_var = "id", biomarker_var = "antigen_iso"),     antigen_isos = c("HlyE_IgG", "HlyE_IgA"), curve_params = structure(list(        y1 = c(63.4814181747235, 163.892554798255, 288.162064918977,         153.628916105779, 431.819308863086, 143.455366610532,         30.5523468510691, 53.3082414171847, 160.248521898273,         557.057672849571, 524.750289905931, 957.183975655821,         30.7757024960889, 696.885680249965, 41.3351250925557,         49.696304772494, 248.45349985843, 284.238616409023, 319.258088721929,         55.1973083368717, 9.85590602194967, 142.48248989252,         58.569079122574, 125.103468262329, 23.5781354491369,         154.134822731266, 195.272088240338, 512.685734719332,         15.2392432670215, 367.675521980743, 138.597844051858,         94.054499427184, 17.7002608315275, 420.52048494015, 34.6297552188594,         117.116522186164, 670.782356315421, 139.74457140291,         350.389335470867, 280.773133101888, 524.805205039604,         341.875466132432, 21.7419806650667, 807.973583062943,         54.648570130211, 161.138851467563, 62.4626331974805,         345.012549159694, 13.5263909774471, 287.580231401241,         404.641741222692, 329.62595529365, 127.460080376226,         119.698649885564, 47.6743496395363, 53.3427557637164,         4.86255586452024, 93.6129541009416, 217.262169334863,         55.1129621346739, 35.0985002310083, 118.566555662152,         1098.49716393263, 54.1079875304336, 58.9629133595119,         135.952389154747, 42.1086528807225, 260.595389430244,         34.7560167599678, 257.405675946662, 179.060328314574,         153.694462686078, 24.6668810561169, 49.4445183661751,         574.894674464689, 105.684016008491, 16.5510503302757,         60.2041812390575, 11.0453659443594, 462.717062144179,         329.450591149241, 382.551153691391, 134.232955969569,         239.461396134792, 14.3719137242382, 264.590126446701,         40.6802550135182, 57.5144029967135, 88.2747290650814,         430.319125620458, 10.6130205814949, 558.11559324373,         32.0135413405647, 24.0384935132398, 38.0504292985456,         77.8837709668701, 126.820466076145, 159.918844849729,         69.8997358428869, 262.882882182499, 354.532775056227,         412.676482549134, 114.923006872123, 120.873465623906,         132.600241082824, 103.468186139286, 447.006072553703,         1085.2206115645, 33.5210558211561, 253.133868952307,         130.757054116255, 66.9218148382036, 374.582473744296,         139.536904862498, 39.3988388717891, 96.9165448765317,         179.919538402879, 1493.88241008644, 14.8008942224083,         404.839466898054, 532.707197471499, 46.4587352210946,         24.4858328538187, 49.4642193897299, 95.0227133002871,         93.9690584818973, 260.176071822173, 141.989238850102,         39.9773724835592, 782.954380644115, 29.4107109741128,         516.645589158684, 211.572242009836, 1822.2130872245,         52.509809995604, 13.151025358795, 19.3857150727236, 35.2409166828083,         80.9687188700813, 177.173315875654, 26.3645555024978,         117.225011423899, 87.0261479522951, 126.950596752532,         27.2083356730063, 83.8209553654887, 65.4624630013765,         60.5914219180724, 50.5312183443017, 220.403393063456,         57.8814946339193, 321.339218295469, 112.331004619454,         214.114844173551, 46.2663050252167, 253.087688470067,         75.8950087172145, 364.707374884755, 21.9277543373803,         342.0606126627, 72.882860491943, 41.3783165219195, 44.191172297619,         74.7393669070667, 50.8282443896275, 265.44382427441,         228.249964982196, 150.22058930907, 137.733518575792,         282.081550999412, 225.506262502798, 31.1882936999509,         163.565411792614, 182.312164883095, 468.092763349791,         26.948324983628, 47.0627893374153, 265.595229664979,         11.1704463639473, 172.041567739628, 161.068053764295,         432.516403298317, 89.2366348763595, 93.0637657208957,         867.380852022527, 402.754319315464, 17.8792209899537,         288.905665377172, 21.9097654634741, 48.8445971425635,         21.8483966121941, 183.574641774107, 42.8450651248465,         392.445793775464, 11.3934815053904, 323.545778561326,         120.315053250724, 164.404167510907, 249.551023456624,         41.4554723971584), alpha = c(0.000580750108201723, 0.00457307241085562,         0.000459482362967931, 0.000920576917699969, 0.000277272439231206,         0.00293646148218711, 0.00126803096870046, 0.000185056384174521,         0.00139768218010029, 0.000505440617097379, 0.0002940737416904,         0.000654669519525582, 0.00459319232964828, 0.000887688202447749,         0.00233809215626583, 0.000209870475682425, 4.67002985675705e-05,         0.00104003594908331, 0.000448299061620713, 0.000207443820402452,         0.00184390439817354, 0.000541774876402651, 0.000165952326601835,         0.00322191298857545, 0.000801608679478142, 0.00117192216751816,         0.00212248013302342, 0.000976984276211827, 0.00206185338337752,         0.000384579510907555, 0.000520896079311473, 0.0010065164190521,         0.00110520990806848, 0.000183194889839954, 0.000503901371757611,         0.000228263475932955, 0.00194228531845488, 0.000171501250472504,         0.000369695819464055, 0.00104853639157351, 0.000697205747962488,         0.000311385146127225, 0.000354295233312761, 0.00124573445413704,         0.000423939415887257, 0.000801256049243512, 0.00120645375923901,         0.00237756125361664, 0.00123608024365872, 0.000358006290675028,         0.000494140830213741, 0.000156002269928014, 0.00245664175975264,         0.000841503824065256, 0.00193807512902148, 0.000497129501207083,         0.0054332923652717, 0.00127310051929336, 0.000551300789665394,         0.00063403271748031, 0.00174555975196488, 0.000377066120482253,         0.000483383608036441, 0.000338217966168421, 0.00219483150170641,         0.00140321394275048, 0.00108547239461618, 0.00102792815759423,         0.00349137373455354, 0.000352559907428524, 0.00162066292577739,         0.00279808506815644, 0.0019601191828733, 0.000330547561933239,         0.000355786402556068, 0.000438855656341692, 0.00124273584887326,         0.000552763674223956, 0.00186976972015624, 0.000333069293847281,         0.00110379614438794, 0.00146362360047569, 0.000139360006251701,         0.000397180389049223, 0.00114211896576467, 0.00100256857212561,         0.00109121884614163, 0.00126819582344172, 0.000987731162476162,         0.00128987495424607, 0.00282261928684757, 0.000374791757339121,         0.00113865219626825, 0.000532967559301844, 0.000513638935226,         0.000359459156219103, 0.000275859190991226, 0.000354699360187289,         0.0026713535164984, 0.00279347203459326, 0.000128197309556073,         0.00596927380308968, 0.00705294513320615, 0.00202300752718538,         0.00162192753599336, 0.00037424870310045, 0.00026754841964219,         0.000730493664235532, 0.00153178738489586, 0.00325308988648181,         0.000483857556168969, 0.000588331496127102, 0.000298345006855044,         0.00112620692068216, 0.00168651670227233, 0.000204975520121039,         0.00193362267850561, 0.00142648771405165, 0.00134562054578092,         0.000208857156789975, 0.00062536817604273, 0.00133456484375557,         0.00938643234422778, 0.000937138007517452, 0.000608514388479188,         0.00264210323257104, 0.00039208182606313, 0.0033181097851919,         0.00185906798532453, 0.000907043634868749, 0.00115294008901915,         0.000245596650454221, 0.00108389960019551, 0.00201213285019601,         0.00190943097530246, 0.000408867260930471, 0.000791381820969645,         0.00201466973548936, 0.00171498662024685, 0.00140127884303538,         0.00100416392412402, 0.000519205915058319, 0.00167112203653147,         0.000838815315735768, 0.000721650217325436, 0.000505339539082098,         0.000909450488173269, 0.00121178596919844, 0.000826864269791345,         0.00292640667315851, 0.000304137699570372, 0.00113983701054901,         0.000389736378485822, 0.00298474787407907, 0.000841259823338093,         0.00368059814723494, 0.000421837664169578, 0.00123986844219214,         0.000760625071667979, 0.00148555030442975, 0.000832930793734274,         0.000300251530379406, 0.00191957782487084, 0.000394689964996692,         0.00267988874066063, 0.00153195589077585, 0.00122125043306422,         0.000373008257684936, 0.00134318717545753, 0.00069797309624943,         0.000766596477494754, 0.000245039778346658, 0.000872995636865741,         0.00712119308128448, 4.98002775826703e-05, 0.000948299668165794,         0.000521991110458308, 0.00125413398046824, 0.000780668464134002,         0.002049580788824, 5.11867422438249e-05, 0.00561075492929095,         0.000808839543407316, 0.000713355973905716, 0.00085947398602869,         0.0021832078345804, 0.00180421443138756, 0.00058909093348889,         0.000599760589426105, 0.00135466647345608, 0.00143236264326595,         0.000494961785196012, 0.000608033851949211, 0.00451563256019836,         0.00337223348145991, 0.000362658114803346, 8.99092256955004e-05,         0.000361336825504191, 0.000266811004105529, 0.00308982754595651        ), r = c(1.74503804112011, 1.17416885836092, 2.66300631906146,         1.297476514491, 1.61055432802302, 1.13118326027234, 1.87013752944932,         2.12891387733085, 1.3974520424861, 1.97493702481855,         2.26045907519679, 1.73951774586965, 1.54839140734894,         1.28946525686499, 1.47712107228344, 1.60869619166956,         8.49637822070033, 1.22786543992899, 1.74930137992105,         2.09051460553557, 1.25152824238038, 1.6479440039637,         1.87839842610789, 1.1688001176233, 1.27395812849041,         1.42241923931779, 1.60775344339114, 1.22367302575217,         1.13545307616552, 1.4528247924265, 1.83738947009373,         1.38484985799107, 1.33445598250036, 2.21809845578665,         2.87854398365401, 2.95589130494591, 1.37775176705415,         1.78770312896554, 3.42523396464135, 1.21637549019504,         1.48417977755664, 2.43501743648731, 1.36792322142982,         1.47888442241868, 1.51608025120621, 1.19711459871732,         2.80191321923623, 1.11128063778834, 1.29719404599601,         1.56958115730821, 1.40645421035182, 2.02055071677547,         1.58553635830427, 1.31900452223899, 1.22922069635161,         1.65815487698244, 1.55892554941562, 1.26151595080374,         1.37707355668328, 1.74441842726501, 1.20180972115458,         1.44365502023906, 2.38510091527588, 1.88284644056197,         1.60597957384943, 1.34867326455727, 1.62070645876627,         1.31154857933272, 2.12645665211127, 1.65075568439601,         1.48036454239674, 1.19982166204518, 2.00237147751504,         1.70487485680804, 1.8706623806063, 1.67166807264746,         1.92194113718141, 1.61005143847568, 1.35004974710959,         2.25841123408653, 1.43335501102913, 1.55790365098244,         4.72154907642204, 2.58282464355033, 1.4444685411938,         1.1934183150664, 2.03619132144836, 1.50070658907823,         1.70535433471468, 1.30762000737011, 1.35345821652327,         2.11594435320902, 1.93343793924938, 1.40636886392934,         1.64717599789262, 3.07065710721104, 1.64913472189956,         1.56811466820133, 1.3743054888023, 1.32153142343595,         1.75972177367333, 1.0321394831503, 1.19619621962556,         1.12522664793909, 2.54761339632466, 1.65664548682786,         1.84384875557502, 1.37448370062602, 1.3397178020426,         1.10640954472882, 2.19064717043815, 1.57332473216103,         1.69127262231831, 1.27894797055983, 1.43991225166163,         3.51828219391506, 1.46683565487797, 1.34968938149023,         1.47588549995483, 1.77726961029852, 1.98675568807676,         1.24408225496149, 1.17673214459276, 1.24982115428208,         1.93255112801193, 1.27076878883401, 2.15156743349361,         1.09741062383505, 1.56474555183983, 1.26556384404366,         1.49417874265552, 1.86821491243254, 1.6685112860065,         1.13876686280646, 3.89157113577646, 2.26881576887324,         1.98350653159569, 1.11306753318267, 1.30383514829675,         1.12087842385809, 1.33146601283398, 1.93446154806597,         1.37443801640134, 1.5782031508516, 2.36840523534241,         1.90640359775196, 1.71299918615934, 1.43614740844928,         2.37312340146821, 1.16058369630921, 1.93968496829668,         1.20625335236676, 1.24209864788451, 1.11020753598308,         1.2758632519492, 1.15080012417002, 2.14686026015812,         1.27879180342719, 2.22148215191972, 1.28974327693987,         1.42731834708861, 2.32334947011738, 1.61955363733894,         1.47492898090976, 1.5350794705679, 1.2441188700458, 1.23133925719503,         1.99038276427008, 1.51776282933421, 1.31925025031428,         1.61304802616839, 1.31460732660273, 1.42961019701532,         1.03565806252171, 4.80132617777742, 1.38736960114353,         1.55863107325236, 1.52666057929324, 2.28991200129968,         1.15325838364272, 6.2383156195126, 1.09899855746642,         1.65867700079462, 1.5005684131288, 1.95624425662808,         1.15457745212576, 1.46923995980707, 1.39171137447131,         1.65402676714906, 1.26867038104759, 2.67952161518617,         1.98222310689708, 2.32213915915393, 1.09182873618516,         2.14613567346721, 1.79173305163522, 2.56558099827198,         1.82633749444038, 2.24145862650252, 1.14228545396753),         antigen_iso = structure(c(1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L,         2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L, 1L, 2L), levels = c("HlyE_IgA",         "HlyE_IgG", "LPS_IgA", "LPS_IgG", "Vi_IgG"), class = "factor")), row.names = c(NA,     -200L), class = c("curve_params", "tbl_df", "tbl", "data.frame"    ), antigen_isos = structure(1:5, levels = c("HlyE_IgA", "HlyE_IgG",     "LPS_IgA", "LPS_IgG", "Vi_IgG"), class = "factor"), biomarker_var = "antigen_iso"),     noise_params = structure(list(nu = c(2.59560775795, 2.3600514143    ), eps = c(0.2793208827, 0.1456403573), y.low = c(0.508452309,     1.594575648), y.high = c(5e+06, 5e+06), antigen_iso = c("HlyE_IgA",     "HlyE_IgG")), row.names = c(NA, -2L), class = c("noise_params",     "tbl_df", "tbl", "data.frame"), antigen_isos = c("HlyE_IgA",     "HlyE_IgG", "LPS_IgA", "LPS_IgG")), lambda_start = 0.1, antigen_isos = c("HlyE_IgG",     "HlyE_IgA"), build_graph = FALSE, print_graph = FALSE, verbose = FALSE): formal argument "antigen_isos" matched by multiple actual arguments
#> Timing stopped at: 0.002 0 0.002

est2sum <- summary(est2)
#> Error: object 'est2' not found

autoplot(est2sum, "catchment")
#> Error: object 'est2sum' not found