Campylobacter Delft Response Parameters Data for Model 3

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

Usage

campylobacterDelftParams3

Format

A list of three dataframes:

IgA

A dataframe containing 4000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 4000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 4000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in campylobacterDelftParams3
lapply(campylobacterDelftParams3, head)
#> $IgA
#>          y1       alpha yb r        y0        mu1        t1
#> 1 0.9790816 0.005617736  0 1 0.5968037 0.16152280 3.0647487
#> 2 0.8561027 0.005686470  0 1 0.6490095 0.14953802 1.8519904
#> 3 0.8115014 0.005653234  0 1 0.3162894 0.20908627 4.5064094
#> 4 1.6246947 0.007071109  0 1 0.7882382 0.29381913 2.4616328
#> 5 1.6912089 0.007544257  0 1 0.5962426 0.25043491 4.1629632
#> 6 7.4984300 0.011753968  0 1 7.3277274 0.06639833 0.3468196
#> 
#> $IgM
#>          y1       alpha yb r        y0       mu1        t1
#> 1 1.4153120 0.004264498  0 1 0.6213528 0.2079336  3.958986
#> 2 0.6276850 0.002730359  0 1 0.3697257 0.1366566  3.873045
#> 3 3.4163114 0.008804830  0 1 0.2768008 0.2385225 10.535773
#> 4 1.8461903 0.005448317  0 1 0.9193840 0.3195337  2.181853
#> 5 0.7790359 0.002590766  0 1 0.5470029 0.0860775  4.107961
#> 6 0.3165559 0.001623372  0 1 0.2296537 0.0809985  3.962139
#> 
#> $IgG
#>          y1       alpha yb r       y0       mu1       t1
#> 1 12.012193 0.003507050  0 1 1.140502 0.3664422 6.425171
#> 2 11.361843 0.001877301  0 1 1.578391 0.3105795 6.355391
#> 3  6.644749 0.001918942  0 1 1.138631 0.3945647 4.470751
#> 4 12.757673 0.001004286  0 1 2.809073 0.2565113 5.899462
#> 5 10.514934 0.001570312  0 1 1.277037 0.6905252 3.053117
#> 6  8.421454 0.003307627  0 1 1.181497 0.3235032 6.071037
#>