Campylobacter Delft Response Parameters Data for Model 4

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

Usage

campylobacterDelftParams4

Format

A list of three dataframes:

IgA

A dataframe containing 3000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 3000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 3000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in campylobacterDelftParams4
lapply(campylobacterDelftParams4, head)
#> $IgA
#>          y1        alpha yb        r         y0       mu1        t1
#> 1 2.7471594 0.0235920904  0 2.773443 0.06452524 0.2737697 13.702269
#> 2 0.3361973 0.0001511561  0 1.006182 0.13241137 0.6774068  1.375517
#> 3 1.3402570 0.0072306458  0 1.640734 0.12395023 0.3883815  6.129892
#> 4 7.8335091 0.0028906244  0 2.768401 0.06512216 0.3482461 13.754356
#> 5 0.3069594 0.0029306951  0 1.050954 0.07870162 0.4218624  3.226293
#> 6 1.6315826 0.0044291436  0 1.740029 0.12264517 0.5924527  4.368299
#> 
#> $IgM
#>          y1        alpha yb        r         y0       mu1       t1
#> 1 3.0218848 0.0068170529  0 1.073372 0.09900034 0.3980767 8.587573
#> 2 1.4605254 0.0004345318  0 1.042603 0.07462244 0.4741715 6.272224
#> 3 2.1067968 0.0043837064  0 1.068606 0.10507351 0.4748134 6.314615
#> 4 1.1187181 0.0019154929  0 1.071340 0.09971466 0.3124635 7.737306
#> 5 0.4222472 0.0005651563  0 1.000181 0.12935880 0.5040532 2.346977
#> 6 4.5818356 0.0158849155  0 1.071558 0.10262796 0.6777377 5.605037
#> 
#> $IgG
#>          y1        alpha yb        r         y0       mu1       t1
#> 1 27.861427 0.0051607511  0 1.019014 0.07791154 0.7296875 8.057456
#> 2  3.306758 0.0011913102  0 1.099056 0.05077868 0.8435221 4.950963
#> 3 18.791581 0.0006869688  0 1.010531 0.09212040 0.6116469 8.694670
#> 4  5.866064 0.0006929407  0 1.019739 0.07694126 0.5232645 8.282421
#> 5 14.364930 0.0035222413  0 1.092096 0.05110497 0.9445758 5.969519
#> 6  5.790477 0.0003697673  0 1.010602 0.09127834 0.7696968 5.391807
#>