Campylobacter SSI Response Parameters Data for Model 2

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type

Usage

campylobacterSSIParams2

Format

A list of three dataframes:

IgA

A dataframe containing 3000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 3000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 3000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in campylobacterSSIParams2
lapply(campylobacterSSIParams2, head)
#> $IgA
#>          y1       alpha yb        r y0 mu1 t1
#> 1 0.3790329 0.013723334  0 1.180966 NA  NA  0
#> 2 1.2994481 0.005642420  0 1.212107 NA  NA  0
#> 3 9.4465428 0.011623566  0 1.205368 NA  NA  0
#> 4 1.0094504 0.027030054  0 1.114516 NA  NA  0
#> 5 0.7431767 0.004296025  0 1.210028 NA  NA  0
#> 6 0.5895802 0.005579473  0 1.219622 NA  NA  0
#> 
#> $IgM
#>          y1       alpha yb        r y0 mu1 t1
#> 1 2.0649209 0.011219203  0 1.163448 NA  NA  0
#> 2 0.5802861 0.001512339  0 1.140606 NA  NA  0
#> 3 1.0067090 0.016613231  0 1.197256 NA  NA  0
#> 4 0.6119739 0.013465649  0 1.144594 NA  NA  0
#> 5 0.6514105 0.002957626  0 1.142666 NA  NA  0
#> 6 0.3944259 0.021869072  0 1.136722 NA  NA  0
#> 
#> $IgG
#>          y1        alpha yb        r y0 mu1 t1
#> 1 1.3482386 0.0025760745  0 1.160319 NA  NA  0
#> 2 0.7507566 0.0009904948  0 1.163093 NA  NA  0
#> 3 1.5366781 0.0027693686  0 1.169187 NA  NA  0
#> 4 1.3857968 0.0060489554  0 1.142833 NA  NA  0
#> 5 1.1860646 0.0023857957  0 1.155256 NA  NA  0
#> 6 1.6407579 0.0053201985  0 1.148251 NA  NA  0
#>