Campylobacter SSI Response Parameters Data for Model 4

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

Usage

campylobacterSSIParams4

Format

A list of three dataframes:

IgA

A dataframe containing 3000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 3000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 3000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in campylobacterSSIParams4
lapply(campylobacterSSIParams4, head)
#> $IgA
#>          y1       alpha yb        r        y0        mu1       t1
#> 1 0.3790329 0.013723334  0 1.180966 0.3084470 0.04388666 4.695573
#> 2 1.2994481 0.005642420  0 1.212107 0.8601382 0.12999444 3.173996
#> 3 9.4465428 0.011623566  0 1.205368 1.0658161 0.23740385 9.190702
#> 4 1.0094504 0.027030054  0 1.114516 0.9711994 0.03547790 1.088834
#> 5 0.7431767 0.004296025  0 1.210028 0.6803783 0.04747750 1.859510
#> 6 0.5895802 0.005579473  0 1.219622 0.4211322 0.13267245 2.536050
#> 
#> $IgM
#>          y1       alpha yb        r        y0        mu1        t1
#> 1 2.0649209 0.011219203  0 1.163448 1.3425809 0.09497971 4.5325273
#> 2 0.5802861 0.001512339  0 1.140606 0.4028595 0.10778318 3.3858110
#> 3 1.0067090 0.016613231  0 1.197256 0.8758258 0.15130671 0.9204797
#> 4 0.6119739 0.013465649  0 1.144594 0.4820747 0.07959144 2.9976911
#> 5 0.6514105 0.002957626  0 1.142666 0.6132294 0.02129935 2.8358109
#> 6 0.3944259 0.021869072  0 1.136722 0.3087485 0.04117147 5.9483978
#> 
#> $IgG
#>          y1        alpha yb        r        y0        mu1       t1
#> 1 1.3482386 0.0025760745  0 1.160319 0.7311131 0.09686797 6.317734
#> 2 0.7507566 0.0009904948  0 1.163093 0.5792900 0.05599269 4.630574
#> 3 1.5366781 0.0027693686  0 1.169187 0.4715269 0.30069910 3.928852
#> 4 1.3857968 0.0060489554  0 1.142833 0.8390058 0.20711698 2.422848
#> 5 1.1860646 0.0023857957  0 1.155256 0.9270293 0.06708116 3.673325
#> 6 1.6407579 0.0053201985  0 1.148251 1.2162005 0.12496860 2.396015
#>