Age specific seroincidence function This function models seroincidence using maximum likelihood estimation; that is, it finds the value of the seroincidence parameter which maximizes the likelihood (i.e., joint probability) of the data.

Age specific seroincidence function This function models seroincidence using maximum likelihood estimation; that is, it finds the value of the seroincidence parameter which maximizes the likelihood (i.e., joint probability) of the data.

Usage

est.incidence(
  dpop,
  dmcmc,
  noise_params,
  c.age = NULL,
  antigen_isos = unique(dpop$antigen_iso),
  start = 0.1,
  iterlim = 100,
  coverage = 0.95,
  verbose = FALSE,
  stepmax = 1,
  ...
)

Arguments

dpop

cross-sectional population data

dmcmc

mcmc samples from distribution of longitudinal decay curve parameters

noise_params

a data.frame() containing columns nu, etc. specifying conditional noise parameters

c.age

age category

antigen_isos

antigen isotypes: a character() vector of one or more antigen isotype names, which should match the values of the antigen_iso column in the dpop input argument

start

starting value for incidence rate

iterlim

an integer(), which provides an upper limit on the number of computational iterations used to search for the maximum likelihood estimate of incidence (passed to stats::nlm()).

coverage

desired confidence interval coverage probability

verbose

logical: if TRUE, print verbose log information to console

stepmax

a numeric(), which limits how aggressively the stats::nlm() algorithm searches for the maximum likelihood estimate of incidence. If this function output an infinite standard error estimate, consider reducing this parameter.

...

Arguments passed on to stats::nlm

typsize

an estimate of the size of each parameter at the minimum.

fscale

an estimate of the size of f at the minimum.

print.level

this argument determines the level of printing which is done during the minimization process. The default value of 0 means that no printing occurs, a value of 1 means that initial and final details are printed and a value of 2 means that full tracing information is printed.

ndigit

the number of significant digits in the function f.

gradtol

a positive scalar giving the tolerance at which the scaled gradient is considered close enough to zero to terminate the algorithm. The scaled gradient is a measure of the relative change in f in each direction p[i] divided by the relative change in p[i].

steptol

A positive scalar providing the minimum allowable relative step length.

check.analyticals

a logical scalar specifying whether the analytic gradients and Hessians, if they are supplied, should be checked against numerical derivatives at the initial parameter values. This can help detect incorrectly formulated gradients or Hessians.

Value

A data.frame() containing the following:

est.start: the starting guess for incidence rate

ageCat: the age category we are analyzing

incidence.rate: the estimated incidence rate, per person year

CI.lwr: lower limit of confidence interval for incidence rate

CI.upr: upper limit of confidence interval for incidence rate

coverage: coverage probability

neg.llik: negative log-likelihood

iterations: the number of iterations used