Estimate Seroincidence

Function to estimate seroincidences based on cross-section serology data and longitudinal response model.

Usage

est.incidence.by(
  data,
  curve_params,
  noise_params,
  strata = "",
  curve_strata_varnames = strata,
  noise_strata_varnames = strata,
  antigen_isos = unique(pull(data, "antigen_iso")),
  lambda.start = 1/365.25,
  numCores = 1L,
  verbose = FALSE,
  ...
)

Arguments

data

Data frame with cross-sectional serology data per antibody and age, and additional columns to identify possible strata.

curve_params

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

noise_params

a list() (or data.frame(), or tibble()) containing noise parameters

strata

Character vector of stratum-defining variables. Values must be variable names in data. Default = "".

curve_strata_varnames

A subset of strata. Values must be variable names in curve_params. Default = "".

noise_strata_varnames

A subset of strata. Values must be variable names in noise_params. Default = "".

antigen_isos

Character vector with one or more antibody names. Values must match data

lambda.start

starting guess for incidence rate, in years/event.

numCores

Number of processor cores to use for calculations when computing by strata. If set to more than 1 and package parallel is available, then the computations are executed in parallel. Default = 1L.

verbose

logical: if TRUE, print verbose log information to console

...

Arguments passed on to .optNll, stats::nlm

stepmax

a positive scalar which gives the maximum allowable scaled step length. stepmax is used to prevent steps which would cause the optimization function to overflow, to prevent the algorithm from leaving the area of interest in parameter space, or to detect divergence in the algorithm. stepmax would be chosen small enough to prevent the first two of these occurrences, but should be larger than any anticipated reasonable step.

typsize

an estimate of the size of each parameter at the minimum.

fscale

an estimate of the size of f at the minimum.

print.level

this argument determines the level of printing which is done during the minimization process. The default value of 0 means that no printing occurs, a value of 1 means that initial and final details are printed and a value of 2 means that full tracing information is printed.

ndigit

the number of significant digits in the function f.

gradtol

a positive scalar giving the tolerance at which the scaled gradient is considered close enough to zero to terminate the algorithm. The scaled gradient is a measure of the relative change in f in each direction p[i] divided by the relative change in p[i].

steptol

A positive scalar providing the minimum allowable relative step length.

iterlim

a positive integer specifying the maximum number of iterations to be performed before the program is terminated.

check.analyticals

a logical scalar specifying whether the analytic gradients and Hessians, if they are supplied, should be checked against numerical derivatives at the initial parameter values. This can help detect incorrectly formulated gradients or Hessians.

Value

A set of lambda estimates for each strata.