Graph an antibody decay curve model

Graph an antibody decay curve model

Usage

graph.decay.curves(
  curve_params,
  verbose = FALSE,
  alpha = 0.4,
  n_curves = 100,
  log_x = FALSE,
  rows_to_graph = sample.int(n = nrow(curve_params), size = min(n_curves,
    nrow(curve_params)), replace = FALSE),
  xlim = c(10^-1, 10^3.1),
  ...
)

Arguments

curve_params

data.frame() of curve parameters (one or more MCMC samples)

verbose

verbose output

alpha

(passed to ggplot2::geom_function()) how transparent the curves should be:

0 = fully transparent (invisible)

1 = fully opaque (no transparency)

n_curves

how many curves to plot (see details).

log_x

should the x-axis be on a logarithmic scale (TRUE) or linear scale (FALSE, default)?

rows_to_graph

which rows of curve_params to plot (overrides n_curves).

xlim

range of x values to graph

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Value

a ggplot2::ggplot() object

Details

n_curves and rows_to_graph

In most cases, curve_params will contain too many rows of MCMC samples for all of these samples to be plotted at once.

Setting the n_curves argument to a value smaller than the number of rows in curve_params will cause this function to randomly select (without replacement) a subset of n_curves rows to graph.

Setting n_curves larger than the number of rows in ` will result all curves being plotted.

If the user directly specifies the rows_to_graph argument, then n_curves has no effect.

Examples

if (FALSE) {
curve_params = readRDS(url("https://osf.io/download/rtw5k/"))
plot1 = graph.curve.params(curve_params)
print(plot1)
}