Load noise parameters

Load noise parameters

Usage

load_noise_params(file_path, antigen_isos = NULL)

Arguments

file_path

path to an RDS file containing biologic and measurement noise of antibody decay curve parameters y.low, eps, nu, and y.high, stored as a data.frame() or tibble::tbl_df

antigen_isos

character() vector of antigen isotypes to be used in analyses

Value

a noise object (a tibble::tbl_df with extra attribute antigen_isos)

Examples

noise = load_noise_params("https://osf.io/download//hqy4v/")
print(noise)
#> # A tibble: 16 × 6
#>    antigen_iso Country    y.low   eps    nu  y.high
#>    <chr>       <fct>      <dbl> <dbl> <dbl>   <dbl>
#>  1 HlyE_IgA    Bangladesh 0.376 0.280  2.60 5000000
#>  2 HlyE_IgA    Ghana      0.179 0.240  2.60 5000000
#>  3 HlyE_IgA    Nepal      0.853 0.238  2.60 5000000
#>  4 HlyE_IgA    Pakistan   0.508 0.279  2.60 5000000
#>  5 HlyE_IgG    Bangladesh 0.787 0.306  2.36 5000000
#>  6 HlyE_IgG    Ghana      0.645 0.164  2.36 5000000
#>  7 HlyE_IgG    Nepal      1.89  0.128  2.36 5000000
#>  8 HlyE_IgG    Pakistan   1.59  0.146  2.36 5000000
#>  9 LPS_IgA     Bangladesh 0.660 0.299  2.14 5000000
#> 10 LPS_IgA     Ghana      0.861 0.163  2.14 5000000
#> 11 LPS_IgA     Nepal      1.79  0.115  2.14 5000000
#> 12 LPS_IgA     Pakistan   5.13  0.246  2.14 5000000
#> 13 LPS_IgG     Bangladesh 0.992 0.298  3.24 5000000
#> 14 LPS_IgG     Ghana      0.885 0.195  3.24 5000000
#> 15 LPS_IgG     Nepal      0.647 0.179  3.24 5000000
#> 16 LPS_IgG     Pakistan   4.84  0.273  3.24 5000000