Salmonella SSI Response Parameters Data for Model 2

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

Usage

salmonellaSSIParams2

Format

A list of three dataframes:

IgA

A dataframe containing 3000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 3000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 3000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in salmonellaSSIParams2
lapply(salmonellaSSIParams2, head)
#> $IgA
#>          y1       alpha yb        r y0 mu1 t1
#> 1 0.1645894 0.008301086  0 1.020289 NA  NA  0
#> 2 0.4967038 0.007181714  0 1.021778 NA  NA  0
#> 3 2.5087166 0.001178859  0 1.017500 NA  NA  0
#> 4 1.8760033 0.018867951  0 1.020985 NA  NA  0
#> 5 0.2381400 0.003545378  0 1.018420 NA  NA  0
#> 6 0.1075755 0.001709320  0 1.017080 NA  NA  0
#> 
#> $IgM
#>          y1        alpha yb        r y0 mu1 t1
#> 1 0.6325777 0.0028957411  0 1.047555 NA  NA  0
#> 2 1.4144767 0.0328629331  0 1.039616 NA  NA  0
#> 3 1.2642730 0.0008913089  0 1.051353 NA  NA  0
#> 4 1.0383240 0.0024629578  0 1.056098 NA  NA  0
#> 5 1.9768815 0.0031380379  0 1.052206 NA  NA  0
#> 6 1.0602443 0.0140293204  0 1.049931 NA  NA  0
#> 
#> $IgG
#>          y1       alpha yb        r y0 mu1 t1
#> 1 0.5630924 0.003808514  0 1.017216 NA  NA  0
#> 2 0.9084731 0.004009900  0 1.021487 NA  NA  0
#> 3 1.3531416 0.001163992  0 1.013681 NA  NA  0
#> 4 0.7098593 0.004429380  0 1.021024 NA  NA  0
#> 5 0.9199583 0.002791685  0 1.020375 NA  NA  0
#> 6 8.4187094 0.003258891  0 1.017908 NA  NA  0
#>