Salmonella SSI Response Parameters Data for Model 4

List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.

Usage

salmonellaSSIParams4

Format

A list of three dataframes:

IgA

A dataframe containing 3000 rows with 7 parameters for IgA antibody.

IgM

A dataframe containing 3000 rows with 7 parameters for IgM antibody.

IgG

A dataframe containing 3000 rows with 7 parameters for IgG antibody.

Examples

# Show first rows of every dataframe contained in salmonellaSSIParams4
lapply(salmonellaSSIParams4, head)
#> $IgA
#>          y1       alpha yb        r         y0        mu1        t1
#> 1 0.1645894 0.008301086  0 1.020289 0.11973340 0.12222920 2.6031943
#> 2 0.4967038 0.007181714  0 1.021778 0.39052803 0.14802164 1.6247227
#> 3 2.5087166 0.001178859  0 1.017500 2.37886174 0.17727577 0.2998108
#> 4 1.8760033 0.018867951  0 1.020985 1.81306536 0.15521542 0.2198533
#> 5 0.2381400 0.003545378  0 1.018420 0.21870973 0.11699959 0.7274678
#> 6 0.1075755 0.001709320  0 1.017080 0.09313794 0.08910887 1.6172527
#> 
#> $IgM
#>          y1        alpha yb        r        y0        mu1        t1
#> 1 0.6325777 0.0028957411  0 1.047555 0.5102927 0.07773161 2.7635934
#> 2 1.4144767 0.0328629331  0 1.039616 0.5753705 0.20969035 4.2896625
#> 3 1.2642730 0.0008913089  0 1.051353 1.1804183 0.07957789 0.8624049
#> 4 1.0383240 0.0024629578  0 1.056098 1.0077723 0.03253768 0.9178773
#> 5 1.9768815 0.0031380379  0 1.052206 1.2086867 0.11864577 4.1466813
#> 6 1.0602443 0.0140293204  0 1.049931 0.8761034 0.05988377 3.1856792
#> 
#> $IgG
#>          y1       alpha yb        r        y0        mu1        t1
#> 1 0.5630924 0.003808514  0 1.017216 0.5040835 0.11686141 0.9472916
#> 2 0.9084731 0.004009900  0 1.021487 0.7985884 0.25997719 0.4958881
#> 3 1.3531416 0.001163992  0 1.013681 0.7592344 0.17166853 3.3662183
#> 4 0.7098593 0.004429380  0 1.021024 0.6843716 0.06179318 0.5917449
#> 5 0.9199583 0.002791685  0 1.020375 0.8494778 0.16093737 0.4952642
#> 6 8.4187094 0.003258891  0 1.017908 3.5463722 1.59327615 0.5426124
#>