collect cross-sectional data

output: (age, y(t) set)

Usage

simcs.tinf(
  lambda,
  n.smpl,
  age.rng,
  age.fx = NA,
  antigen_isos,
  n.mc = 0,
  renew.params = FALSE,
  ...
)

Arguments

lambda

seroconversion rate (in events/person-day)

n.smpl

number of samples n.smpl (= nr of simulated records)

age.rng

age range to use for simulating data, in days

age.fx

age.fx for parameter sample (age.fx = NA for age at infection)

antigen_isos

Character vector with one or more antibody names. Values must match curve_params.

n.mc

when n.mc is in 1:4000 a fixed posterior sample is used

when n.mc = 0 a random sample is chosen

renew.params

renew.params = TRUE generates a new parameter set for each infection

renew.params = FALSE keeps the one selected at birth, but updates baseline y0

...

arguments passed to simresp.tinf()

Value

an array()