Split data by stratum

Split biomarker data, decay curve parameters, and noise parameters to prepare for stratified incidence estimation.

Usage

stratify_data(
  data,
  curve_params,
  noise_params,
  strata_varnames = "",
  curve_strata_varnames = NULL,
  noise_strata_varnames = NULL,
  antigen_isos = data %>% attr("antigen_isos")
)

Arguments

curve_params

a data.frame() containing MCMC samples of parameters from the Bayesian posterior distribution of a longitudinal decay curve model. The parameter columns must be named:

antigen_iso: a character() vector indicating antigen-isotype combinations

iter: an integer() vector indicating MCMC sampling iterations

y0: baseline antibody level at $t=0$ ($y(t=0)$)

y1: antibody peak level (ELISA units)

t1: duration of infection

alpha: antibody decay rate (1/days for the current longitudinal parameter sets)

r: shape factor of antibody decay

noise_params

a data.frame() (or tibble::tibble()) containing the following variables, specifying noise parameters for each antigen isotype:

antigen_iso: antigen isotype whose noise parameters are being specified on each row

nu: biological noise

eps: measurement noise

y.low: lower limit of detection for the current antigen isotype

y.high: upper limit of detection for the current antigen isotype

strata_varnames

character() vector of names of variables in data to stratify by

curve_strata_varnames

A subset of strata. Values must be variable names in curve_params. Default = "".

noise_strata_varnames

A subset of strata. Values must be variable names in noise_params. Default = "".

antigen_isos

Character vector with one or more antibody names. Values must match pop_data

Value

a "biomarker_data_and_params.list" object (a list with extra attributes "strata", "antigen_isos", etc)

Examples

if (FALSE) { # \dontrun{
library(dplyr)

xs_data <- load_pop_data("https://osf.io/download//n6cp3/")

curve <- load_curve_params("https://osf.io/download/rtw5k/") %>%
  filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG")) %>%
  slice(1:100, .by = antigen_iso) # Reduce dataset for the purposes of this example

noise <- load_noise_params("https://osf.io/download//hqy4v/")

stratified_data =
  stratify_data(
   data = xs_data,
   curve_params = curve,
   noise_params = noise,
   strata_varnames = "catchment",
   curve_strata_varnames = NULL,
   noise_strata_varnames = NULL
   )
} # }