Summary Method for "seroincidence.ests" Objects

Calculate seroincidence from output of the seroincidence calculator est.incidence.by().

Usage

# S3 method for seroincidence.ests
summary(
  object,
  confidence_level = 0.95,
  showDeviance = TRUE,
  showConvergence = FALSE,
  ...
)

Arguments

object

A dataframe containing output of function est.incidence.by().

confidence_level

desired confidence interval coverage probability

showDeviance

Logical flag (FALSE/TRUE) for reporting deviance (-2*log(likelihood) at estimated seroincidence. Default = TRUE.

showConvergence

Logical flag (FALSE/TRUE) for reporting convergence (see help for optim() for details). Default = FALSE.

...

Additional arguments affecting the summary produced.

Value

A list with the following items:

Results a dplyr::tibble() with columns:

incidence.rate maximum likelihood estimate of lambda (seroincidence)

CI.lwr lower confidence bound for lambda

CI.upr upper confidence bound for lambda

Deviance (included if showDeviance = TRUE) Negative log likelihood (NLL) at estimated (maximum likelihood) lambda)

nlm.exit.code (included if showConvergence = TRUE) Convergence information returned by stats::nlm()

Antibodies Character vector with names of input antibodies used in est.incidence.by()

Strata Character with names of strata used in est.incidence.by()

Examples

if (FALSE) {
# estimate seroincidence
seroincidence <- est.incidence.by(...)

# calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)
}