Load noise parameters

Load noise parameters

Usage

load_noise_params(file_path, antigen_isos = NULL)

Arguments

file_path

path to an RDS file containing biologic and measurement noise of antibody decay curve parameters y.low, eps, nu, and y.high, stored as a data.frame() or tibble::tbl_df

antigen_isos

character() vector of antigen isotypes to be used in analyses

Value

a noise object (a tibble::tbl_df with extra attribute antigen_isos)

Examples

noise <- load_noise_params(serocalculator_example("example_noise_params.rds"))
print(noise)
#> # A tibble: 4 × 8
#>       X antigen_iso Country  y.low   eps    nu  y.high Lab  
#>   <int> <chr>       <chr>    <dbl> <dbl> <dbl>   <dbl> <chr>
#> 1     1 HlyE_IgA    Pakistan 0.508 0.279  2.60 5000000 AKU  
#> 2     2 HlyE_IgG    Pakistan 1.59  0.146  2.36 5000000 AKU  
#> 3     3 LPS_IgA     Pakistan 5.13  0.246  2.14 5000000 AKU  
#> 4     4 LPS_IgG     Pakistan 4.84  0.273  3.24 5000000 AKU