simulate antibody kinetics of y over a time interval

simulate antibody kinetics of y over a time interval

Usage

simresp.tinf(
  lambda,
  t.end,
  age.fx,
  antigen_isos,
  n.mc = 0,
  renew.params,
  predpar,
  ...
)

Arguments

lambda

seroconversion rate (1/days),

t.end

end of time interval (beginning is time 0) in days(?)

age.fx

parameter estimates for fixed age (age.fx in years) or not. when age.fx = NA then age at infection is used.

antigen_isos

antigen isotypes

n.mc

a posterior sample may be selected (1:4000), or not when n.mc = 0 a posterior sample is chosen at random.

renew.params

At infection, a new parameter sample may be generated (when renew.params = TRUE). Otherwise (when renew.params = FALSE), a sample is generated at birth and kept, but baseline y0 are carried over from prior infections.

predpar

an array() with dimensions named:

antigen_iso

parameter

obs

...

Arguments passed on to row_longitudinal_parameter, ab, mk_baseline

age

age at infection

nmc

mcmc sample to use

npar

number of parameters

t

age at infection?

par

parameters

kab

index for which row of antibody baseline limits to read from blims

n

number of observations

blims

range of possible baseline antibody levels

Value

This function returns a list() with:

t = times (in days, birth at day 0),

b = bacteria level, for each antibody signal (not used; probably meaningless),

y = antibody level, for each antibody signal

smp = whether an infection involves a big jump or a small jump

t.inf = times when infections have occurred.